+++ bcbio/structural/shared.py
@@ -17,7 +17,25 @@ from bcbio.bam import callable
 from bcbio.pipeline import datadict as dd
 from bcbio.pipeline import shared
 from bcbio.provenance import do
-from bcbio.variation import population
+from bcbio.variation import effects, population, vcfutils
+
+# ## Finalizing samples
+
+def finalize_sv(orig_vcf, data, items):
+    """Finalize structural variants, adding effects and splitting if needed.
+    """
+    paired = vcfutils.get_paired(items)
+    # For paired/somatic, attach combined calls to tumor sample
+    if paired:
+        sample_vcf = orig_vcf if paired.tumor_name == dd.get_sample_name(data) else None
+    else:
+        sample_vcf = "%s-%s.vcf.gz" % (utils.splitext_plus(orig_vcf)[0], dd.get_sample_name(data))
+        sample_vcf = vcfutils.select_sample(orig_vcf, dd.get_sample_name(data), sample_vcf, data["config"])
+    if sample_vcf:
+        effects_vcf, _ = effects.add_to_vcf(sample_vcf, data, "snpeff")
+    else:
+        effects_vcf = None
+    return effects_vcf or sample_vcf
 
 # ## Case/control